% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overshadow.R
\name{check_overshadow}
\alias{check_overshadow}
\title{Check Overshadow}
\usage{
check_overshadow(choice_sets)
}
\arguments{
\item{choice_sets}{An object of the choiceset class made by one of the DCE methods.}
}
\value{
A matrix of logicals indicating which if any card for a given row is Pareto optimal.
}
\description{
Check Overshadow
}
\examples{
See step 7 of the practical vingette
}
