% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_question_frame.R
\name{construct_question_frame}
\alias{construct_question_frame}
\title{Convert from choice_sets to a question dataframe}
\usage{
construct_question_frame(choice_sets, af, m,
  randomise_choice_sets = TRUE)
}
\arguments{
\item{choice_sets}{The choice sets list generated from one of the methods. (See Step 6 of the tutorial)}

\item{m}{The number of alternatives in each choice set}

\item{randomise_choice_sets}{A binary variable indicating if the order of the choice sets should be randomised. Some methods create choice sets which have a systematic order. Randomising the order of the choice sets does not change the alternatives within the choice sets. It simply rearranges the choice_set object in a random manner.}
}
\value{
a data.frame object
}
\description{
Convert from choice_sets to a question dataframe
}
\examples{
See Step 8 of tutorial
}
