% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dce_efficiency.R
\name{dce_efficiency}
\alias{dce_efficiency}
\title{Efficiency Measures for Discrete Choice Experiments}
\usage{
dce_efficiency(augmented_full_factorial, choice_sets)
}
\arguments{
\item{augmented_full_factorial}{The level augmented full factorial. See tutorial step 2.}

\item{choice_sets}{A list of choice sets generated by one of the methods used to convert from fractional factorial designs.}
}
\value{
a list of named output.
}
\description{
Efficiency Measures for Discrete Choice Experiments
}
\examples{
# See Step 8 of the Practical Introduction to ExpertChoice vignette.
}
\references{
Street, D.J., Burgess, L. and Louviere, J.J., 2005. Quick and easy choice sets: constructing optimal and nearly optimal stated choice experiments. International Journal of Research in Marketing, 22(4), pp.459-470.
}
