% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dce_effiency.R
\name{dce_effiency}
\alias{dce_effiency}
\title{Efficacy Measures for Discrete Choice Experiments}
\usage{
dce_effiency(augmented_full_factorial, choice_sets)
}
\arguments{
\item{augmented_full_factorial}{The level augmented full factorial. See tutorial step 2.}

\item{choice_sets}{A list of choice sets generated by one of the methods used to convert from fractional factorial designs.}

\item{m}{Number of alternatives in each choice set}
}
\value{
a list of named output.
}
\description{
Efficacy Measures for Discrete Choice Experiments
}
\examples{
See tutorial step 8
}
