% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo_method.R
\name{modulo_method}
\alias{modulo_method}
\title{Modulo Method - Described by Street et al.}
\usage{
modulo_method(fractional_fatorial, generators)
}
\arguments{
\item{fractional_fatorial}{The usual.}

\item{generators}{a list of generators}
}
\value{
a choiceset list.
}
\description{
Modulo Method - Described by Street et al.
}
\examples{
# See step 6 of the Practical Introduction to ExpertChoice vignette.
}
