% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_design.R
\name{search_design}
\alias{search_design}
\title{Search Full Factorial for Fractional Factorial Design}
\usage{
search_design(full_factorial, fractional_factorial_design)
}
\arguments{
\item{full_factorial}{a `data.table` generated by the `full_factorial` function}

\item{fractional_factorial_design}{a means of creating a fractional design using either orthogonal arrays or Federov. See the tutorial for examples.}
}
\value{
a `data.frame` with only the rows of your chosen fractional fractorial design.
}
\description{
Returns a consistent fractional factorial design from the input fractional factorial design. The key advantage of this function is that it ensures factors are coded and enchances the attributes of the output.
}
\examples{
# The use of this function depends on what the input to the argument fractional_factorial_design will be.
# See Step 4 of the tutorial (on the GitHub repo) for more detials
}
