% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_4521_64-data.R
\docType{data}
\name{silver_4521_64}
\alias{silver_4521_64}
\title{An ortogonal array with 64 rows designed for the silver expert example in the tutorial.}
\format{An MIP orthogonal array.}
\source{
Stephens, J. Desiginng Silver Resarch.
}
\usage{
silver_4521_64
}
\description{
An orthogonal array of resolution IV.
}
\details{
None more.
}
\keyword{datasets}
