% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stephens_pairing.R
\name{stephens_pairing}
\alias{stephens_pairing}
\title{Stephens Pairing}
\usage{
stephens_pairing(fractional_factorial)
}
\arguments{
\item{fractional_factorial}{A fractional factorial design.}
}
\value{
Returns a list which includes the choice_set object.
}
\description{
Currently only returns a discrete choice experiment with two alternatives.
Does not allow one alternative to overshadow the other.
In the optimised object, each row is a choice set. (These are marked by the row names).
In the p1 column is the row from the fractional_factorial object which is matched with the p2 row from the fractional factorial_factorial.
(Hence they are "pairs".)
}
\examples{

}
