% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overshadow.R
\name{check_overshadow}
\alias{check_overshadow}
\title{Check Overshadow - Pareto Dominate Solutions}
\usage{
check_overshadow(choice_sets)
}
\arguments{
\item{choice_sets}{An object of the choiceset class made by one of the DCE methods.}
}
\value{
A matrix of logicals indicating which if any card for a given row is Pareto dominate.
}
\description{
Check Overshadow - Pareto Dominate Solutions
}
\examples{
#See Step 7 of the Practical Introduction to ExpertChoice Vignette.
}
