% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional_factorial_efficiency.R
\name{fractional_factorial_efficiency}
\alias{fractional_factorial_efficiency}
\title{Fractional Factorial Design Efficiency}
\usage{
fractional_factorial_efficiency(formula, searched_fractional_factorial)
}
\arguments{
\item{formula}{A specification, in formula form, of the desired effects sought to be estimated.}

\item{searched_fractional_factorial}{a fractional factorial generated as the result of a `search_design`.}
}
\description{
Fractional Factorial Design Efficiency
}
\examples{
See tutorial for worked examples.
}
\references{
Kuhfeld, W. F. Marketing Research Methods in SAS Experimental Design, Choice, Conjoint, and Graphical Techniques 2010.
}
